chat-app/requirements.txt
langchain==0.2.6
langchain-community==0.2.6
langchain-core==0.2.10
langchain-openai==0.1.13
langchain-text-splitters==0.2.2
python-dotenv
streamlit==1.36.0
fastapi==0.111.0
fastapi-cli==0.0.4
uvicorn==0.30.1
sse_starlette==2.1.2
bs4==0.0.2
pypdf==4.2.0
chromadb==0.5.3
pydantic==2.8.0
flashrank==0.2.6
PyPDF2==3.0.1
streamlit-chat==0.1.1
html2text==2020.1.16
arize-phoenix==4.12.0
arize-phoenix-evals==0.14.1
